Add '-fPIC' to gcc CFLAGS and CXXFLAGS
This is necessary to compile, likely due to libcc1.so always being
generated (even if we try to disable it and only go for its static
equivalent), which links in libstdc++.a. -fPIC needs to be set on both
CFLAGS and CXXFLAGS, just doing it on CXXFLAGS failed to compile.